task

+++ print_dig.c
@@ -48,7 +48,7 @@ int print_p_dig(long n)
  * Return: size of output
 */
 
-int print_HEX(unsigned int n)
+int print_HEX(long n)
 {
 	int ctr = 0;
 	int r, i = 0, j;
